Move Slice into its own package
